--- src/Components/Bottombar.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import { BrowserRouter } from 'react-router-dom'
-
-export default function Bottombar() {
-    return (
-        <div>
-            Bottombar
-        </div>
-    )
-}
