+++ src/Components/Form.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import {useEffect} from 'react'
-import moment from 'moment'
+import {postData, putData, utcToLocal, localToUtc ,checkInterviewValidations} from "../Utils"
+import {reflectFormErrors} from "../AuthUtils"
 
 export default function Form(props) {
     let titleRef = null;
@@ -9,68 +10,21 @@ export default function Form(props) {
     let participantsRef = null;
     let interviewerRef = null;
 
-    const setData = () => { // this function runs automatically and initializes the form values
+    const setData = () => { 
         titleRef.value = props.title ? props.title : ""
-        if(props.starttime){
-            let localvalue =  moment.utc(props.starttime).local().format();
-            localvalue = localvalue.substr(0, localvalue.indexOf("+"));
-            starttimeRef.value = localvalue;
-        } 
-        if(props.endtime){
-            let localvalue =  moment.utc(props.endtime).local().format();
-            localvalue = localvalue.substr(0, localvalue.indexOf("+"));
-            endtimeRef.value = localvalue;
-        } 
+        starttimeRef.value = props.starttime ? utcToLocal(props.starttime) : ""
+        endtimeRef.value = props.endtime ? utcToLocal(props.endtime) : "" 
         interviewerRef.value = props.members ? props.members.interviewer.id : ""
     }
-    const postData = async (url = '', data = {}) => {
-        const response = await fetch(url, {
-            method: 'POST',
-            mode: 'cors',
-            cache: 'no-cache',
-            credentials: 'same-origin',
-            headers: {
-                'Content-Type': 'application/json',
-                'Authorization': `Bearer ${
-                    localStorage.getItem("token")
-                }`
-            },
-            redirect: 'follow',
-            referrerPolicy: 'no-referrer',
-            body: JSON.stringify(data)
-        });
-        const jsondata = await response.json();
-        return jsondata;
-    }
 
-
-    const putData = async (url = '', data = {}) => {
-        const response = await fetch(url, {
-            method: 'PUT',
-            mode: 'cors',
-            cache: 'no-cache',
-            credentials: 'same-origin',
-            headers: {
-                'Content-Type': 'application/json',
-                'Authorization': `Bearer ${
-                    localStorage.getItem("token")
-                }`
-            },
-            redirect: 'follow',
-            referrerPolicy: 'no-referrer',
-            body: JSON.stringify(data)
-        });
-        const jsondata = await response.json();
-        return jsondata;
-    }
-    const getData = () => {
+    const collectFormData = () => {
         let interview = {}
         interview.id = props.id
         interview.title = titleRef.value
         interview.starttime = starttimeRef.value
         interview.endtime = endtimeRef.value
-        interview.starttime = new Date(interview.starttime).toUTCString();
-        interview.endtime = new Date(interview.endtime).toUTCString();
+        interview.starttime = localToUtc(interview.starttime);
+        interview.endtime = localToUtc(interview.endtime);;
         interview.interviewer = interviewerRef.value
         interview.participants = []
         let options = participantsRef.options
@@ -84,71 +38,16 @@ export default function Form(props) {
     const onClose = (e) => {
         props.changeEditState ? props.changeEditState() : props.changeCreateState();
     }
-    const checkInterviewValidations = ({
-        title,
-        starttime,
-        endtime,
-        interviewer,
-        participants
-    }) => {
-        if(!starttime)
-            console.log("$$$$$$$$$$$$$$$$$$$$$$$4")
-        console.log("%%%%%%%%%%%%%%%%%%%%%%%%%%%%%", title)
-        const errors = {}
-        if (title === "") 
-            errors.title = "please provide a title"
-        
-        if (endtime === "") 
-            errors.endtime = "please provide an end time"
-        
-        if (starttime === "") 
-            errors.starttime = "please provide a start time"
-        
-        if (!interviewer) 
-            errors.interviewer = "please select an interviewer"
-        
-        if (starttime !== "" && new Date(starttime) <= new Date()) 
-            errors.starttime = "this ain't a time machine!!"
-        
-        if (title !== "" && title.length < 3) 
-            errors.title = "provide a valid title with minimum length of 3";
-        
-        if (starttime !== "" && endtime && starttime >= endtime) 
-            errors.starttime = "starttime must be less than endtime";
-        
-        if (participants.length == 0) 
-            errors.participants = "atleast one participant must be selected";
-        
-        return errors;
-    }
-    const reflectFormErrors = (errors) => {
-        document.getElementById("title").innerText = errors.title ? errors.title : ""
-        document.getElementById("starttime").innerText = errors.starttime ? errors.starttime : ""
-        document.getElementById("endtime").innerText = errors.endtime ? errors.endtime : ""
-        document.getElementById("interviewer").innerText = errors.interviewer ? errors.interviewer : ""
-        document.getElementById("participants").innerText = errors.participants ? errors.participants : ""
-    }
+    
 
     const onClick = (e) => {
-        // e.preventDefault();
-        // const formData = getData();
-        // const errors = checkInterviewValidations(formData);
-        // if(Object.keys(errors).length !== 0)
-        //     reflectFormErrors(errors);
-        // else{
-        //     props.onEdit ? props.onEdit(e, formData) : props.onCreate(e, formData);
-        //     onClose();
-        // }
-
-        // on click I need to create interview or edit interview and then if everything goes right, I will call props.onEdit and props.onCreate
         e.preventDefault();
-        const formData = getData();
+        const formData = collectFormData();
         const errors = checkInterviewValidations(formData);
         if (Object.keys(errors).length !== 0) 
             reflectFormErrors(errors);
          else {
             props.id ? editInterview(formData) : createInterview(formData);
-            // props.onEdit ? props.onEdit(e, formData) : props.onCreate(e, formData);
         }
     }
