--- src/Pages/Create.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Create() {
-    return (
-        <div>
-            Create
-        </div>
-    )
-}
