--- src/Pages/Edit.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Edit() {
-    return (
-        <div>
-            Edit
-        </div>
-    )
-}
