+++ src/Pages/InterviewList.js
@@ -2,87 +2,14 @@ import React, {useState, useEffect} from 'react'
 import Interview from '../Components/Interview'
 import Form from '../Components/Form'
 import { useHistory } from "react-router-dom";
+import {checkLoginStatus} from "../AuthUtils"
+import { getData, deleteData} from "../Utils"
 
 export default function InterviewList(props) {
     let history = useHistory();
     const [Interviews, setInterviews] = useState([]);
     const [Users, setUsers] = useState([]);
 
-    const checkLoginStatus = async (url = '') => {
-        const response = await fetch('http://localhost:4000/loggedIn', {
-            method: 'GET',
-            mode: 'cors',
-            cache: 'no-cache',
-            credentials: 'same-origin',
-            headers: {
-                'Content-Type': 'application/json',
-                'Authorization': `Bearer ${
-                    localStorage.getItem("token")
-                }`
-            },
-            redirect: 'follow',
-            referrerPolicy: 'no-referrer'
-        });
-        const jsondata = await response.json();
-        console.log(jsondata);
-        if(jsondata.id){
-            return true;
-        }
-        return false;
-    }
-
-    const getData = async(url = '')=> {
-        const response = await fetch(url, {
-          method: 'GET',
-          mode: 'cors',       
-          cache: 'no-cache', 
-          credentials: 'same-origin', 
-          headers: {
-            'Content-Type': 'application/json',
-            'Authorization': `Bearer ${localStorage.getItem("token")}`
-          },
-          redirect: 'follow', 
-          referrerPolicy: 'no-referrer'
-        });
-        const jsondata = await response.json();
-        return jsondata;
-    }
-    const postData = async(url = '', data = {})=> {
-        const response = await fetch(url, {
-          method: 'POST',
-          mode: 'cors',       
-          cache: 'no-cache', 
-          credentials: 'same-origin', 
-          headers: {
-            'Content-Type': 'application/json',
-            'Authorization': `Bearer ${localStorage.getItem("token")}`
-          },
-          redirect: 'follow', 
-          referrerPolicy: 'no-referrer', 
-          body: JSON.stringify(data)
-        });
-        const jsondata = await response.json();
-        return jsondata;
-    }
-    const deleteData = async(url = '')=> {
-        const response = await fetch(url, {
-          method: 'DELETE',
-          mode: 'cors',       
-          cache: 'no-cache', 
-          credentials: 'same-origin', 
-          headers: {
-            'Content-Type': 'application/json',
-            'Authorization': `Bearer ${localStorage.getItem("token")}`
-          },
-          redirect: 'follow', 
-          referrerPolicy: 'no-referrer', 
-        });
-        const jsondata = await response.json();
-        return jsondata;
-    }
-    // const [createState, setCreateState] = useState(false);
-    // this function fetches data and sets initial state
-    // right now lets mimic the functionality
     const fetchUsers = async () => {
         const nextUsers = Object.assign([], Users);
         let fetchedUsers = await getData("http://localhost:4000/users"); 
@@ -99,6 +26,7 @@ export default function InterviewList(props) {
         });
         return Promise.all(promises);
     }
+
     const fetchInterviewsData = async() => {
         const nextInterviews = [];
         let interviews = await getData("http://localhost:4000/interviews");
@@ -107,9 +35,9 @@ export default function InterviewList(props) {
         fetchedData.forEach((interview) => {
             nextInterviews.push(interview);
         })
-        console.log("next interviews are ", nextInterviews);
         setInterviews(nextInterviews);
     };
+
     useEffect( () => {
         checkLoginStatus().then((res)=>{
             if(res){
